refactor: extract diagnostic project analysis
Closes diagnostic-query-project-analysis-extraction.
