+++ ai_docs/backlog.md
@@ -3,7 +3,7 @@
 <!-- purpose: Open work only. Slim-index format — triage in the table, implementation detail in items/<id>.md. Sync rows on ship. -->
 <!-- scope: in-repo -->
 
-**updated_at:** 2026-08-31T13:51:05Z
+**updated_at:** 2026-08-31T15:40:06Z
 
 ## Agent contract
 
@@ -156,7 +156,7 @@
 | `workspace-readiness-fixture-ready-state-isolation` | Low | — | **Automate readiness order proof** — keep isolated scenarios and add a bounded repeated runner that executes the class in deliberately different orders. [type: test correctness] [source: 2026-08-24 dependency review] | S | items/workspace-readiness-fixture-ready-state-isolation.md |
 | `code-pattern-analyzer-responsibility-split` | Low | — | **Split CodePatternAnalyzer responsibilities** — extract reflection-usage scanning from semantic-query parsing and search behind compatibility-preserving contracts. [type: refactor] [source: 2026-08-27 diagnostic scan review] | M | items/code-pattern-analyzer-responsibility-split.md |
 | `ci-merge-gate-publish-gate-shape-divergence` | Low | — | **Unsharded coverage exists only on a developer box** — add a scheduled unsharded Windows `verify-release.ps1` leg so pre-tag unsharded coverage does not depend on a workstation being healthy. [type: ci] [source: 2026-08-28 release-cut] | M | items/ci-merge-gate-publish-gate-shape-divergence.md |
-| `diagnostic-query-project-analysis-extraction` | Low | — | **Extract diagnostic project analysis** — move analyzer severity resolution and per-project collection out of the 529-line query/cache service; preserve config precedence, totals, and cancellation. [type: refactor] [source: 2026-08-31 adjacent review] | M | items/diagnostic-query-project-analysis-extraction.md |
+| `diagnostic-analyzer-descriptor-enumeration-failure-isolation` | Low | — | **Isolate analyzer descriptor enumeration failures** — make Error-only fast-path probing degrade conservatively when third-party analyzer enumeration throws; preserve cancellation and emit safe diagnostics. [type: reliability] [source: adjacent review] | M | items/diagnostic-analyzer-descriptor-enumeration-failure-isolation.md |
 
 ## Defer
