+++ changelog.d/diagnostic-query-project-analysis-extraction.md
@@ -1,0 +1,5 @@
+---
+category: Changed
+---
+
+- **Changed:** Diagnostic queries now delegate per-project compiler, generator, analyzer, and effective-severity work to a focused collaborator with direct precedence coverage. Closes `diagnostic-query-project-analysis-extraction`.
